+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added support for parsing Vi-Cell XR txt files
 - Add Electrophoresis BENCHLING/2024/06 schema.
 - Added github enforcement that CHANGELOG.md is updated
 - Added test files Agilent TapeStation Analysis