+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+<<<<<<< mabtech_apex
+- Add Mabtech Apex adapter
+=======
 - Added support for parsing Vi-Cell XR txt files
 - Add Electrophoresis BENCHLING/2024/06 schema.
+>>>>>>> main
 - Added github enforcement that CHANGELOG.md is updated
 - Added test files Agilent TapeStation Analysis
 - Added requirements for remaining parsers to /docs