+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Fixed
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Security
+
+## [0.1.35] - 2024-06-24
+
+### Added
 - Add Mabtech Apex adapter
 - Added support for parsing Vi-Cell XR txt files
 - Add Electrophoresis BENCHLING/2024/06 schema.